+++ docs/whatsnew.md
@@ -3,6 +3,7 @@
 *V1.2.0*
 - Object: the objectDescriptor can now contain a field 'LinkDescriptor'
 - add Object.evmask to change link descriptor event mask
+- add Object.hysteresis to read/write event hysteresis
 - Variable.valueChanged now accepts a callback with signature ( value, Object )
 - new Object: new Member userName.
 - Variable/Task/Module: use name as userName if 'CD' is used for defining plc object